requirements.txt
# Updated requirements with security fixes - January 2025
# Core web framework and dependencies
flask>=3.0.3
requests>=2.32.0
urllib3>=2.2.2
certifi>=2024.7.4
jinja2>=3.1.4
werkzeug>=3.0.3
itsdangerous>=2.2.0
markupsafe>=2.1.5

# Meraki API and utilities
meraki>=1.47.0
python-dotenv>=1.0.1
click>=8.1.7
colorama>=0.4.6

# Testing framework
pytest>=8.0.0

# Additional security-focused packages
cryptography>=42.0.0
